Add cricket country images. Move image mapping to new file
